Job opening and list components somewhat fleshed out
